# MAINT: Speedup field access by removing unneeded safety checks