#5548 causes a big slowdown in the specific case of accessing/assigning fields of structured arrays with large (eg 1kb) dtypes. Oops! I noticed this while investigating #1984. The problem is #5548 implemented an algorithm which checks view safety byte-by-byte (slow) under the assumption that most dtypes are only a small number of bytes. That's bad if the dtype is very large, as in the test script of #1984.

One fix is to speed the safety-check algorithm up, but it also turns out that the checks aren't needed in many cases. For example, during field indexing they aren't needed because we're merely viewing fields that already exist, which is therefore safe. So I've bypassed the safety checks by rewriting that indexing code (in C) so it avoids using `PyArray_View`. 

I also made corresponding voidtype methods call the ndarray methods to get the same benefits. Incidentally this also enables some functionality that was missing from voidtype relative to ndarray. Eg, you can now do `arr[0][['a', 'b']]` and it will give you a more correct error message. 

Safety checks are still important for get/setfield and for views. But even there we can skip the checks early if we know that the datatypes aren't structured or aren't of object type. 

Timing info for the test script in #1984, run as `./test.py 10000`

Numpy 1.9.2:

    array,  B['a'][:] = A['a']      0.00852 seconds
    array,  B['a']    = A['a']      0.00648 seconds
    scalar, B['a'][:] = A['a']      0.016 seconds
    scalar, B['a']    = A['a']      0.132 seconds

Master:

    array,  B['a'][:] = A['a']      16.4 seconds
    array,  B['a']    = A['a']      16.3 seconds
    scalar, B['a'][:] = A['a']      16.1 seconds
    scalar, B['a']    = A['a']      16.6 seconds

This PR:

    array,  B['a'][:] = A['a']      0.00733 seconds
    array,  B['a']    = A['a']      0.00673 seconds
    scalar, B['a'][:] = A['a']      0.0171 seconds
    scalar, B['a']    = A['a']      0.0151 seconds


Later I'll look into a faster algorithm for the structured safety checks.

I think this PR also fixes #1984, if 0.0151s is fast enough! 

By the way, I think get/setfield are no longer used anywhere internally in numpy.