@seberg 2015-08-19
Do we have to consider this a 1.10 regression?