@ahaldane 2015-08-19
I don't yet know enough about the dev process to answer that, but note it's probably a pretty rare problem:

The slowdown should only be noticeable if you are taking 1000s of views of an array with a huge dtype. I expect that the number of people using huge dtypes (eg, large subarrays) is pretty small, and then the number of people also taking 1000s of views with such a dtype is even smaller.