@ahaldane 2015-10-13
I'll write up a short description of the code changes in a little bit, to help any reviewers.