@seberg 2015-10-13
Not sure you have to worry about it too much, back in the day I wanted to know if we need to make sure it makes it into 1.10. I thought it did not sound like it would be necessary (which apparently was not quite right). After that, never looked at it. If you think it makes reviewing much easier, go ahead I guess.