@ahaldane 2015-10-13
To help anyone reviewing, here is a summary of the code changes:

Previously, `_index_fields` would be given a list of field names and a structured array, and 1. take a view of the array of only those fields, 2. for multi-field indices make a copy of that view, copying only those fields. Step 1 was slow because of the new safety checks for views. Step 2 was needed because views of multiple fields don't work properly, so we have been returning a copy for now.

This PR does essentially the same thing except that step 1 is now written in C to avoid safety checks, and code has been reorganized a little. In total:

 * The code previously in `_index_fields` (1)  is now written in C in `_get_field_view`, so it can 'manually' take a view using `PyArray_NewFromDescr` to avoid the view safety checks. `_get_field_view` also takes over the job of `obj_is_string_or_stringlist` since there was a lot of overlap. So in total `_get_field_view` both checks whether the input index is a field or list of fields, and then takes the view if so.
 * The code for (2) is now a new function `_copy_fields` which performs the copy. (My ultimate goal in the future is to remove step 2, to fix #3641).
 * The view safety checks are now applied only to structured arrays and object arrays, by checking the result of the new function `_may_have_objects`.
 * `_unpack_field` was moved to 'common.c' because it is now used in two places: in the `arraytypes.c` functions for ndarrays, and in `mapping.c` in `_get_field_view`, since both places need to get field info from the fields dict. I think it makes sense to have it in common anyway, since it should be generally useful anytime someone wants to work with structured arrays in numpy code.

Reading over that I'm not sure how clear it is, but if you have any questions let me know.