@pv 2015-10-13
Maybe add a benchmark for it? https://github.com/pv/numpy/commit/1dff685e787db71c727d55b6f5ee76057c3fa4e2