@ahaldane 2015-10-13
Great, I tried adding your commit. Hopefully that is the right way to do it.