@charris 2015-10-13
>Do we have to consider this a 1.10 regression?

Looks like it ;)

This is a pretty extensive refactoring, so it makes me a bit nervous. Someone needs to do a deep review of the code, which I haven't managed yet. Did you ever profile the original to see just where the bottlenecks were?