@ahaldane 2015-10-14
I'm don't think profiling is needed: I know the view safety checks are slow for large datatypes, and I know they are used when indexing fields. So the solution is to (safely) skip the checks here.