@ahaldane 2015-10-18
I am also not against reverting #5548. I knew that the algorithm there was slow, but didn't expect it to be such a problem. Given the extent of the changes needed here, I'm thinking #5548 needs to be given more thought - eg maybe the alternate approach I discussed but decided against would avoid this slowdown.

I just tried reverting it on my local setup, there does not seem to be any problem in doing so. (I just needed one extra small change to account for extra safety features I introduced elsewhere.) Reverting is the easy fix to these speed issues, and I could reintroduce a better version of #5548 for 1.11 or later. Any opinions?