@charris 2015-10-18
@ahaldane If you fix the nit I'll put this in so it can get some testing.