@charris 2015-10-18
Reversion would be the easy fix. I think it comes down to whether of not you would like to take time to rethink the issue  and put up a cleansheet solution or continue to pursue the current fixes.