@ahaldane 2015-10-18
Looking over it all again (and adding a minor tweak), I think merging this is better than reverting #5548. #5548 is probably the right solution, at least until a new dtype system comes out. If not, it's pretty easy to back out of later, and it only introduces a performance problem. 

If you do merge, I've left #6467 open so we can wait for people to report back that their particular case is fixed.