@charris 2015-10-18
OK, let's give it a shot. Thanks Allan.